requirements.txt
pandas
numpy
pyyaml
scikit-learn
xgboost
shap
streamlit
matplotlib
seaborn
joblib
tqdm